+++ components/login/btn/textBtn.jsx
@@ -1,7 +1,7 @@
 import styles from "./textBtn.module.scss";
-const TextBtn = ({ text }) => {
+const TextBtn = ({ text, onClick }) => {
   return (
-    <button type="button" className={styles.textBtn}>
+    <button type="button" className={styles.textBtn} onClick={onClick}>
       <span className={styles.text}>{text}</span>
     </button>
   );
