+++ pages/main.module.scss
@@ -4,7 +4,7 @@
   width: 376px;
   height: 100vh;
   margin: 0 auto;
-  background-color: #f3f3f3; //white로 바꾸기
+  background-color: #fff;
 
   .imageLogo {
     margin: 226px auto 15px;
